split large test suite
